[refactor] better file management for widgets
